+++ src/main/java/com/example/gistcompetitioncnserver/comment/domain/CommentRepository.java
@@ -1,4 +1,4 @@
-package com.example.gistcompetitioncnserver.comment;
+package com.example.gistcompetitioncnserver.comment.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
